+++ src/Symfony/Component/Security/Core/Authentication/AuthenticationTrustResolver.php
@@ -22,23 +22,6 @@ use Symfony\Component\Security\Core\Authentication\Token\TokenInterface;
  */
 class AuthenticationTrustResolver implements AuthenticationTrustResolverInterface
 {
-    private $anonymousClass;
-    private $rememberMeClass;
-
-    public function __construct(?string $anonymousClass = null, ?string $rememberMeClass = null)
-    {
-        $this->anonymousClass = $anonymousClass;
-        $this->rememberMeClass = $rememberMeClass;
-
-        if (null !== $anonymousClass && !is_a($anonymousClass, AnonymousToken::class, true)) {
-            @trigger_error(sprintf('Configuring a custom anonymous token class is deprecated since Symfony 4.2; have the "%s" class extend the "%s" class instead, and remove the "%s" constructor argument.', $anonymousClass, AnonymousToken::class, self::class), E_USER_DEPRECATED);
-        }
-
-        if (null !== $rememberMeClass && !is_a($rememberMeClass, RememberMeToken::class, true)) {
-            @trigger_error(sprintf('Configuring a custom remember me token class is deprecated since Symfony 4.2; have the "%s" class extend the "%s" class instead, and remove the "%s" constructor argument.', $rememberMeClass, RememberMeToken::class, self::class), E_USER_DEPRECATED);
-        }
-    }
-
     /**
      * {@inheritdoc}
      */
@@ -46,10 +29,6 @@ class AuthenticationTrustResolver implements AuthenticationTrustResolverInterfac
     {
         if (null === $token) {
             return false;
-        }
-
-        if (null !== $this->anonymousClass) {
-            return $token instanceof $this->anonymousClass;
         }
 
         return $token instanceof AnonymousToken;
@@ -62,10 +41,6 @@ class AuthenticationTrustResolver implements AuthenticationTrustResolverInterfac
     {
         if (null === $token) {
             return false;
-        }
-
-        if (null !== $this->rememberMeClass) {
-            return $token instanceof $this->rememberMeClass;
         }
 
         return $token instanceof RememberMeToken;